requirements.txt
Flask==2.2.3
flask_mongoengine==1.0.0
flask_wtf==1.1.1
WTForms==3.0.1
pymongo
python-dotenv==0.21.0
